lab3/catalog-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
httpx==0.24.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9 